+++ ex00/main.cpp
@@ -3,10 +3,10 @@
 
 void    test_vector() {
     std::cout << "------ VECTOR ------" << std::endl;
-    Vector<float> u = {2., 3.};
+    Vector<double> u = {2., 3.};
     std::cout << "u = " << std::endl << u << std::endl;
 
-    Vector<float> v = {5., 7.};
+    Vector<double> v = {5., 7.};
     std::cout << "v = " << std::endl << v << "---\n" << std::endl;
 
     std::cout << "u.add(v) = \n";
@@ -28,10 +28,10 @@ void    test_vector() {
 
 void    test_matrix() {
     std::cout << "------ MATRIX ------" << std::endl;
-    Matrix<float> m1 = {{1., 3.}, {2., 4.}};
+    Matrix<double> m1 = {{1., 3.}, {2., 4.}};
     std::cout << "m1 = " << std::endl << m1 << std::endl;
 
-    Matrix<float> m2 = {{7., -2.}, {4., 2.}};
+    Matrix<double> m2 = {{7., -2.}, {4., 2.}};
     std::cout << "m2 = " << std::endl << m2 << "---\n" << std::endl;
     
     std::cout << "m1.add(m2) = \n";
